Issue-9761 - Extract array_empty and array_repeat functions from functions-array subcrate' s kernels and udf containers
